More pure version of Factory Pattern
There was a flaw in the previous factory example. You shouldn't have any awareness of what class will be created by looking at the calling code. In the original example the types were, for the most part, explicitly determined by the caller. For example, the programmer explicitly passing in .UnitedStates creates a UnitedStatesDollar class. In my new example, when looking at the calling code, one has no clue as to what class will be instantiated.

Having said that, I feel like there may be a better way to deal with passing in the file twice but I can't think of one other than using the constructor inside the Switch statement. Thoughts?